requirement.txt
appdirs==1.4.3
asgiref==3.3.1
beautifulsoup4==4.9.3
blis==0.7.4
bs4==0.0.1
CacheControl==0.12.6
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cymem==2.0.5
distlib==0.3.0
distro==1.4.0
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
en-core-web-sm==2.3.1
filelock==3.0.12
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
Markdown==3.3.3
msgpack==0.6.2
murmurhash==1.0.5
mysqlclient==2.0.2
numpy==1.19.4
packaging==20.3
pep517==0.8.2
plac==1.1.3
preshed==3.0.5
progress==1.5
pyparsing==2.4.6
PySocks==1.7.1
pytoml==0.1.21
pytz==2020.4
requests==2.22.0
requests-file==1.5.1
retrying==1.3.3
six==1.14.0
soupsieve==2.1
spacy==2.3.5
sqlparse==0.4.1
srsly==1.0.5
thinc==7.4.5
tldextract==3.1.0
tqdm==4.54.1
urllib3==1.25.8
wasabi==0.8.0
webencodings==0.5.1
